Fixed some z-index issues.
The blue bar that was displayed “under” the page had z-index of -1 and
after the previous change improving text/html “compatibility”, the it
stopped displaying.  Changing it to any other value caused it to be
showed “on top of” the page.  The simple solution was to add z-index
to the div element wrapping the whole of the content of the web
page (this of course required also adding positioning to it).