Improve layout on small screens
If screen is not wide enough, move sidebar to the bottom of the page.
To make this work, disable infinite scrolling and also add Categories
and Archive links to the top menu (which are only visible if sidebar
is at the bottom).

Furthermore, as the screen gets narrower and narrower, remove more
and more eye-candy which consumes horizontal space.

All those changes should dramatically improve the site on mobile
devices.