Drop onums from H1
The reason for of ‘onums’ feature was to make the digits in ‘mina86.com’
title the same height as the rest of the text.  Unfortunately, I’ve just
realised, they actually aren’t the same hight as lower case letters even
if old-style numbers are used.  With that realisation, drop the feature.