Remove white background for the #v element, i.e. the wrap around whole page
Remove white background from the #v element which is a 60-em wide wrapper
around the whole content of the page.  This background resulted with a long
white background.  The bad thing about it was that articles used a border
which resulted in two levels close to each other:

  +---------------------+
  | #v             left |
  | +-----------+  side |
  | | .c        |  bar  |
  | |           |       |
  ⋮ ⋮           ⋮       ⋮

Over time, this started to feel too ‘busy’ to me.  With this change, the
outer border is removed:

   #v             left
   +-----------+  side
   | .c        |  bar
   |           |
   ⋮           ⋮

A gentle box-shadow adds a 3D feeling to the .c elements.  The end result
is that the page now has two layers: the background (with right side bar,
meta information, footer and similar data) and the article content.  This,
I hope, results in a clearer layout.

To match this redesign, the cloud background for the header needed to be
reworked as well.