Stop using Google Drive for news images
Jogger had rather low limit on static file storage which forced me to
move some of the images to Google Drive.  With new hosting this is no
longer an issue so the files can be served from the same server again.
This has an added bonus of working faster than Google Drive.