Fix content type of JavaScript and Atom files
NFS used ‘application/javascript’ for .js files which is arguably the
right thing to do (and technically more correct than ‘text/javascript’)
but, at the same time, the default value of SCRIPT’s TYPE attribute is
is ‘text/javascript’ in HTML5 so I’ll go ahead and go with that instead.

Furthermore, the only thing I use XML for is Atom feeds so go ahead and
set content type of all *.xml files to be that of an Atom feed.