Provide redirects for more types of feeds
Jogger supported various configuration options with its feeds such as
possibility to strip HTML tags, includes excerpts only and choose between
RSS and Atom formats but the new site has this limited to merely one kind
of Atom feed.

Include redirects for all of the Jogger-supported formats to the Atom feed
so that users will get something even if they wanted this served in a bit
different form.

Also mention that in the update post.