Move #b to the top of the page
If a page takes noticeable time to load, having #b at the very end
of the document causes the blue background to suddenly appear on the
screen once the page is loaded.  This leads to bad user experience.
Move #b to the front of the file so this effect is absent.