Change /p/ path to /$year/ instead
So I got more and more unhappy about /p/ directory for all the posts.  Sure
it’s short and sweet but it has its drawbacks.  For one, date when post was
published is kinda useful and maybe useful enough to put in URL.  Full date
still feels too verbose but just the year should be fine.  Furthermore, now
going up a directory from /$year/$post/ to /$year/ leads users to a list of
posts for given year whereas previously /p/ yielded a 404.

This does does add more noise, but hopefully not too much and the pain will
be over soon.