Make some files never-expire
Use content-based naming for some of resource files so that they can be
marked as never-expiring.

This new naming means that the names can no longer be hard-coded in
template and CSS files, or more precisely that there needs to be
a filter which will map the hard-coded names to their dynamic names.

Necessaity of this filter is why this commit moves a lot of logic from
a Makefile into a new Python script.  The script, while copying files,
constructs a mapping between local names to content-based names and
substitutes the names where needed.