Fix leading lines cleanup in PRE; add new lines after <pre> opening tag
Fix the HTML minifier to properly remove leading empty lines in
a PRE element and with that in place, add a new line character after
a <pre> opening tag in multi-line code samples.  This helps since in
multi-line code any alignment between first and second line was so
far problematic.