Use parametric font size consistency; enlarge atricle titles
Replace hard-coded font sizes with the use of constants defined in the
LESS file.  Furthermore, increase size of the article title to make it
more apparent.

While doing all of it, round font sizes to just one digit after the point.
Extra precision is not necessary for anything and in fact may hurt if
browser tries to calculate sub-pixel sizes for the fonts.