Use flexible boxes to keep footer at the bottom of the viewport
Setting body’s background colour to header’s background causes rendering
artefacts in mobile Firefox.  Or rather, when the page is panned, parts
which are not yet rendered are filled with header’s background colour
before they get filled with actual content.  This looks rather bad so
change the body’s background colour to the background colour used in
body text.

However, this causes footer to hang in the middle of the page on very
high viewports when the page is short.  Fix that by using flexible boxes
CSS layout mode.  The result is even better since now the footer stays
the same size and sticks to the bottom of the viewport rather than being
extended.