css: use filter CSS property for grayscale effect on sidebar images
caniuse.com claims over 90% of browsers support filter property (with
notable exception of Internet Explorer, of course) so it’s high time
to use it especially since the effect is just a eye-candy so it’s not
end of the world if 10% of users won’t see it.

With this change, the image with sprites (which this commit renames to
sprites.png which is much more descriptive than t.png) can be reduced
by getting rid of grayscale images.  Adding filter properties mase CSS
file a bit larger but overall result is reduction in size:

	sprites.png  3845 → 2853  Δ -992
	css.css      7997 → 7963  Δ  +34
	css.css.gz   2714 → 2747  Δ  +33