Change font-family settings
1) Make <pre> mono-space again.  In many places proportional font works
fine but there are some where mono-space font is necessary.  Using
different fonts would make it confusing so stick to mono-space in all
<pre> elements.  inline <code> still uses proportional.

2) Use Ubuntu Mono for mono-space and Ubuntu for proportional code
font.  Being of the same family they have somewhat similar looks so it
makes sense to have one of them be for <pre> and another for <code>.

3) Settle on Libre Baskerville for the body font.  Also remove custom
headers font since that just complicates the page without much gain.

4) Import Ubuntu Mono, Ubuntu and Libre Baskerville from Google Fonts so
there’s bigger guarantee that user will use those fonts.

5) Use font-variant for setting font variants (don’t bother with
font-feature-settings or any of the specific font-variant-foo properties)
and while at it turn tabular-nums inside of a table.