Fix sidebar position updats being delyed and simplify JS code
Remove setTimeout from the onscroll and onresize handlers, name some
of the variables better and use Array.prototype.map to shorten the
script.

With the setTimeout call, if user continuously scrols the page (i.e.
lets mouse’s scroll wheel roll), Chrome would delay update of the
position till the scrolling stops.  This sounds good for limiting
computations but it means that sidebar would overshoot or stay out
of the viewing area in certain situations.