Embed sprits in CSS file
Before this change compressed CSS file was 3165 bytes and the sprits
were additional 1868 bytes for total of 5033 bytes.  After this
change, compressed CSS file is 4762 bytes.

Even when comparing uncompressed files, before the change it was
10355 + 1868 for total of 12223 bytes and after the change it’s 12475
bytes for an increase of 252 bytes.  Considering all headers going
back and forth with HTTP requests, even that is probably worth it.

This is a bit of a cheat since Polish flag was moved out of the CSS
file into pages/cma but since that’s the only page that uses it, it is
a reasonable approach.