Replace quote and feed images with SVG versions
Since the new quote image (used in front of blockquote elements) and
feed icon (used as a link to atom link in categories list) are now
vector graphics, they will scale nicely when page is zoomed and render
better on high-DPI displays.

Furthermore, the quote image has a neat shadow.  This change also
improves the red background in psa-chmod post by adding some padding
between the end of text and end of the background.

The feed icon has been replaced with one without the square and
simplified to include the waves only.  Hopefully people using feed
readers will still recognise it and at the same time the new image is
lighter.

As a result of this change, the CSS file shrunk from 4762 to 4658
though uncompressed grew from 12475 to 12628.