Remove .np class and clean CSS more aggressively
This reduces the size of the CSS file:

                  │ Before │ After
    ──────────────┼────────┼───────
     Uncompressed │  14197 │ 13654
     Compressed   │   5350 │  5341

Size of the compressed file shirkns only by nine bytes because all
the optimisation does is removes duplicate rules (i.e. the same rule
repeated separately for different selectors) which is something
compression can deal with to some extent as well.

The optimisations are technically unsafe but work for the page.