Fix make jobserver not being available in child make processes
In Python 3, the default of close_fds argument of subprocess.Popen et
al changed to True.  As a result, jobserver-related file descriptors
which make passes to its children (when -j flag is in effect) are closed
and sub-make processes cannot access the jobserver.  Fix by setting the
argument to False if presence of jobserver is detected.

While at it, if jobserver is not detected, count number of processors
rather than assuming four and run compression with number of jobs twice
the number of processors.