Update rewrite rules and fix some 404s
A few things related to rewriting rules which handle wrong or obsolete
URLs.

1) Fix Atom LINK element to point to /atom rather than non-existent
/atom.xml.  This generated some 404s.

2) Update rewrite rules for Atom files so that more requests are
handled.  In particular, /atom.pl.xml.gz/short/html/ was quite popular
for some reason even though it leads to 404.  Add a rewrite to handle
it and give whoever requsets it a proper result.

3) Stop generating rewrites for obsolete /p/… paths and update such
obsolete link on CMA page to use the new address.  The link sadly lead
to 404 anyway because the generated rewrites were limited in time.

4) Extend handling of old URL to catch broken URLs comming from some
Chinese site which produces things like:
/2010/01/16/ntp-over-http%C2%A0Tambi%C3%A9n

5) Handle ‘-precomposed’ suffix for Apple touch icon files.

6) Remove rewrites for various URLs which haven’t appeared in logs for
over a year or haven’t resulted in a successfull rewrite.