Don’t enforce HTTPS in .htaccess; configure it on the hosting
Enable HTTPS enforcement on NFS such that it does not need to be
specified in .htaccess files.

While a it, drop ‘CSP: upgrade-insecure-requests’ since it’s not
really necessary for anything.  Sure, it prevents resources needed
for the site from being requested over insecure channel, but in
reality none of the resources overwrite the schema to http.