Add ‘sRGB↔XYZ conversion’ article
In doing so, introduce INCLUDE tag which can be used inside of an
article file to include contents of another file.  This is used to
add source code listing to the page without having to keep the code
in HTML.

Also, add the conversion matrix at the end of ‘Calculating sRGB↔XYZ
matrix’ article.  It was a glaring omission.