Fix link to the entry in Atom feed
The rel=self links are used in Atom feeds to point to the feed file itself.
To link to the entry’s version on the website, rel=alternate should be used
instead.  Do that.

While at it, minify the content of the entry in the feed and simplify the
minification code by removing support for stripping base href from URLs.
Instead, strip the base href already in the source files.