Updates to @media print formatting
Firstly, the full URL to the page is already included under the title of
the article so ‘mina86.com’ name at the top does not add any information
and can be removed.

Secondly, since contact block is now just name and email address, print
email in the by-line and get rid of the contacts block at the bottom.

Lastly, add some tweaks to srgb-xyz-conversion article for better
printing.