Minor wording improvements and add image example to sprites-as-bg article
Since the site no longer uses flags and quote icon is done differently
than described in ‘css-sprites-as-background’ article, replace the
live example with a rendered image.  Also add an update note
explaining what has happened.

With that, do various wording changes to aforementioned article as
well as a few others.