Simulate background-attachment: fixed for the top bar image
background-attachment: fixed and background-size: cover do not place
nicely which is why so far the top header used non-fixed attachment.
However, with some JavaScript and explicit background sizing it is
possible to simulate the effect.  The JavaScript code runs only when
window is resized so it’s not computationally intensive either and
effect is smooth when scrolling.