Rework sizes in CSS
Use sizes which are multiples of ⅛.  With browsers using 1rem of 16px
by default using sizes which are fractions with eight in denominator
minimises non-integer pixel dimensions.  This probably doesn’t matter
hugely, but aligning things to pixels on a pixel-based display sounds
like a good idea.

Also increase line-height in the body of the article.