Remove ‘index’ LINKs and make ‘canonical’ relative
There’s already ‘home’ LINK on the page with the same value as ‘index’
so drop the latter.

Furthermore, change ‘canonical’ LINKs to be relative rather than
including the domain.  However user reaches the page they always get
redirected to the main domain so there’s no reason to keep the domain.

This reduces size of the page by over 26 kB.