Fix scrolling to anchors
Because of the fixed position header at the top of the page, whenever
browser scrolls the page to an anchor, the top of the anchor is cut
off.  Fix by adding a JavaScript code which adds an offset to the
scrolling such that the element appears just below the header.