Construct excerpt and full text in advance and store it in Body
Rather than messing around with __parts and constructing excerpt or
full text of the article on demand (when html or __str__ method is
called) generate appropriate stirngs in advance and store them in
__data property.

This simplifies the code somewhat by making it unnecessary to think
about the different length of __parts depending if Body has excerpt
or not.  This will also make it easy to add support for ‘excerpt only’
text in an article.