posts: explicit-isnt-better: two more Rust examples
First of all, implicit () return.  The compiler is quite clear calling
this behaviour implicit, e.g.:

    1 | fn foo() -> u32 { }
      |    ---      ^^^ expected `u32`, found `()`
      |    |
      |    implicitly returns `()` as its body has no tail or `return` expression

Second of all, implicit format! arguments are now in nightly so that
counts too.  The feature is called `format_args_implicits` so it’s
also quite clear that it is an implicit behaviour.

With that also improve some wording in the article.