Remove footer with copyright information
The licensing information gets quite confusing in practice.  Apache
License is incompatible with GPL so I’d like to at least dual-license
the code in the articles.  But then the footer gets quite long indeed.

I could change the license to MIT or some such but I don’t want to
change license to more permissive without looking through all the code
samples on the page.

The reality is that no one will actually read the footer so the
simplest solution is to just remove that information.  It’s easier
to just drop the footer and deal with all situations on case-by-case
basis in case anyone ever asks for license.