js: add (cite) link in byline
Introduce (cite) link to byline which shows a dialogue box with citation
for the article.  Will anyone ever use it? Probably not.  Was it fun to
implement? Yeah, quite fun.

With that also remove some of the old articles which no longer have
much relevance.  This cleanup was prompted by having to go through
all articles to make sure that the titles are all properly annotated
so that BibTeX entry can be generated automatically.