js: don’t delay drop down display and don’t show shade under settings
Firstly, the drop downs fading in and out may look cool but they
introduce unnecessary delay with the use of the site’s interface.
Remove the fading animation for the drop downs.  Fading for the shade
is kept.

Secondly, the settings drop down allows changing appearance of the
website.  However, with the shade active, what exactly the settings do
isn’t obviously apparent and user may be prompted to repetitively hide
and show the drop down to test effect of the settings.  Don’t shade
the website when settings drop down is shown.

Lastly, adjust styles of the backdrop shade shown.  Specifically, add
backdrop-filter which blurs the page and turns it into greyscale.