css: don’t limit floats margins
Up till now, floats were limited to how far to the side they could go.
They could extend no more than 10em to the left or right of the main
text.  Change it so that they can hang completely on the margin if
there’s enough space there for them.