css: multiple style changes
Most notably, reduce chroma of the main background in dark mode.  It
makes it feel ‘darker’ in a way and I think it works better.  With
that, don’t shift hue of the header background any longer but increase
its chroma so it stays somewhat distinct from the main background.

Reduce size of text in H1 and H2 elements though to compensate in a way
increase margins before headers and between entries on index page.

Plus various other smaller changes such as reduction of line-height in
various places.