js: fix pressing Enter on toggle buttons always toggling dark mode
The keydown handler did not take into consideration which of the two
toggles were the target of a Return press resulting in the key press
always toggling dark mode.  Fix it by simulating a click and having
the click handler figure out what needs to be done.

While at it, simplify the click handler by using a single regex to
check element’s class.