posts: switch footnotes style
Use superscripted numbers rather than symbols.  Part of the reason is
that there seems to be no consensus whether the symbols should be put
in superscript or not.  I don’t like the look of dagger and double
dagger without superscript but there are sources that indicate that
this is how it sholud be set.  Switch to numbers to resolve this
ambiguity.