tools: strip back ids on index pages
Paragraphs containing footnotes have ids assigned to them to allow for
footnotes to link back to the paragraph.  If those paragraphs in the
excerpt part of the article their ids must be globally unique to avoid
situation where two such articles appear on the same index page.
Those identifiers are never used though since navigating to the
footnote opens the particle’s page.

Add code which strips those ids from index pages and drops the
requirement for the ids to be globally unique.