data: fix and optimise SVG flags
Most importantly, fix the red diagonal stripes on the UK flag and fix
the SVG so that it renders correctly.  Somehow Firefox didn’t render
it correctly.  I’m not sure whether it was some recent change to the
browser but with some fixes to SVG file Firefox now renders it
correctly.

Secondly, remove the black border around the flags.  The flags are
rendered in such a small size that those borders aren’t visible.

Both files are now slightly smaller so it’s advantages all around.