tpl: add Share on Bluesky, and change Contact
Most importantly, add ‘Share on Bluesky’ link in by-line of all the
posts.  Modify the M font to include a Bluesky logo and use it instead
of rendering the tekt ‘Bluesky’.

Since I’m adding that, also add my Bluesky account in the Contact drop
down.  And since the Bluesky logo already is in the M font, use that
as indicator that the contact is for Bluesky account.

Further refain the Contact list:
- Use ‘CV’ instead of ‘Resume’ to better fit shape of the list.
- Render the contacts on single line when printing.