posts: cma-resources: convert cma page into a post
cma was the only remaining page.  Convert it to a post which gives
the benefit of introducing better discoverability and handling of
features such as citations.

Dating the post was a bit tricky since it was originally a page on
Jogger and that didn’t have date tracked.  Set it to date of the
migration.