posts: use ↩ for footnote back links
Having an explicit back link at the end of the footnote is clearer to
the reader and also the link is introduced at the moment the user most
needs it (after they’re done with the footnote).  Having the footnote
number act as the back link is more cryptic.

And while at it, add role=doc-backlink to all the links which might
help screen readers.